# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  errs.AppError:
    properties:
      code:
        description: |-
          Code 为唯一错误码，由 "DOMAIN.SNAKE_CASE" 格式的 ASCII 字符串构成。
          是前端做国际化映射（i18n key）和告警分组的依据。
        type: string
      details:
        description: |-
          Details 携带字段级校验错误明细，通常用于请求体（DTO）参数校验失败场景。
          如果无字段级错误，该 slice 应为 nil（JSON 序列化时被 omitempty 省略）。
        items:
          $ref: '#/definitions/errs.FieldDetail'
        type: array
      message:
        description: |-
          Message 是面向终端用户的中文提示文案，可直接渲染到 UI。
          要求：简洁、可理解、无敏感堆栈信息。前端根据此字段作为 fallback 文案展示。
        type: string
    type: object
  errs.FieldDetail:
    properties:
      field:
        description: |-
          Field 为请求参数名或 JSON path（如 "title"、"assignee.id"），
          前端可根据此字段将错误文案定位到对应表单控件。
        type: string
      reason:
        description: |-
          Reason 为该字段的具体错误原因（中文），如 "不能超过 255 个字符"、"格式不符合邮箱规范"。
          要求：简洁、无敏感信息、可直接渲染。
        type: string
    type: object
...
  intake.IntakeIssue:
    properties:
      channel_id:
        type: integer
      channel_name:
        type: string
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      linked_entity_id:
        type: integer
      linked_entity_identifier:
        type: string
      linked_entity_type:
        type: string
      name:
        type: string
      priority:
        type: string
      project_id:
        type: integer
      resolved_at:
        type: string
      resolved_by:
        type: integer
      status:
        $ref: '#/definitions/intake.IssueStatus'
      submitter_email:
        type: string
      submitter_name:
        type: string
      tracking_id:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
  intake.IssueStatus:
    enum:
    - open
    - accepted
    - rejected
    - archived
    type: string
    x-enum-varnames:
    - IssueOpen
    - IssueAccepted
    - IssueRejected
    - IssueArchived
...
  issue.Issue:
    properties:
      assignees:
        items:
          type: integer
        type: array
      category:
        type: string
      completed_at:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      depth:
        type: integer
      found_phase:
        type: string
      id:
        type: integer
      identifier:
        type: string
      is_draft:
        type: boolean
      labels:
        items:
          type: integer
        type: array
      modules:
        items:
          type: integer
        type: array
      name:
        type: string
      parent_id:
        type: integer
      point:
        type: integer
      priority:
        $ref: '#/definitions/issue.IssuePriority'
      progress:
        type: integer
      project_id:
        type: integer
      public_id:
        type: string
      sequence_id:
        type: integer
      severity:
        description: 按类型的可选字段
        type: integer
      sort_order:
        type: number
      sprint_id:
        type: integer
      start_date:
        type: string
      state:
        $ref: '#/definitions/issue.State'
      state_id:
        type: integer
      target_date:
        type: string
      type_code:
        $ref: '#/definitions/issue.IssueTypeCode'
      updated_at:
        type: string
      version:
        type: integer
      version_id:
        type: integer
      workspace_id:
        type: integer
    type: object
...
  issue.IssuePriority:
    enum:
    - urgent
    - high
    - medium
    - low
    - none
    type: string
    x-enum-varnames:
    - PriorityUrgent
    - PriorityHigh
    - PriorityMedium
    - PriorityLow
    - PriorityNone
...
  issue.IssueTypeCode:
    enum:
    - epic
    - requirement
    - task
    - defect
    type: string
    x-enum-varnames:
    - TypeEpic
    - TypeRequirement
    - TypeTask
    - TypeDefect
  issue.Label:
    properties:
      code:
        type: string
      color:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      issue_count:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
  issue.Module:
    properties:
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      id:
        type: integer
      lead_id:
        type: integer
      name:
        type: string
      project_id:
        type: integer
      public_id:
        type: string
      sort_order:
        type: number
      start_date:
        type: string
      status:
        type: string
      target_date:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
...
  issue.State:
    properties:
      applicable_types:
        description: 适用的工作项类型：all|epic|requirement|task|defect
        items:
          type: string
        type: array
      color:
        type: string
      created_at:
        type: string
      group:
        $ref: '#/definitions/issue.StateGroup'
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      project_id:
        type: integer
      sequence:
        type: number
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
  issue.StateGroup:
    enum:
    - backlog
    - started
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - GroupBacklog
    - GroupStarted
    - GroupCompleted
    - GroupCancelled
...
host: localhost:8080
info:
  contact: {}
  description: 项目管理平台 Ydsz Plane 的 REST API（需求/任务/缺陷/迭代/版本/自动化/Webhook/知识库/仪表盘/全局搜索/通知等）。
  title: Ydsz Plane API
  version: 1.0.0
paths:
...
  /intake/issues:
    get:
      description: 按渠道/状态分页列出匿名提报工单
      parameters:
      - description: 工单状态
        in: query
        name: status
        type: string
      - description: 渠道 ID
        in: query
        name: channel_id
        type: integer
      - description: 项目 ID
        in: query
        name: project_id
        type: integer
      - default: 50
        description: 每页数
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出收件箱工单
      tags:
      - intake
  /intake/issues/{issue_id}:
    get:
      parameters:
      - description: 工单 ID
        in: path
        name: issue_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/intake.IntakeIssue'
      summary: 获取工单详情
      tags:
      - intake
...
  /issues:
    get:
      description: 跨类型列表，支持稀疏字段集（fields）和 RFC3339 时间范围筛选。
      parameters:
      - description: 状态 ID
        in: query
        name: state_id
        type: integer
      - description: 状态分组 (backlog|started|completed|cancelled)
        in: query
        name: group
        type: string
      - description: 类型 (epic|requirement|task|defect)
        in: query
        name: type
        type: string
      - description: 优先级
        in: query
        name: priority
        type: string
      - description: 父级 ID
        in: query
        name: parent_id
        type: integer
      - description: 名称搜索
        in: query
        name: search
        type: string
      - description: 指派人 ID
        in: query
        name: assignee_id
        type: integer
      - description: 标签 ID
        in: query
        name: label_id
        type: integer
      - description: 模块 ID
        in: query
        name: module_id
        type: integer
      - description: 迭代 ID
        in: query
        name: sprint_id
        type: integer
      - description: 开始日期起 (ISO date)
        in: query
        name: start_date_from
        type: string
      - description: 截止日期止 (ISO date)
        in: query
        name: target_date_to
        type: string
      - description: 最低严重级别
        in: query
        name: severity_from
        type: integer
      - description: 创建时间下限 (RFC3339, e.g. 2025-01-01T00:00:00Z)
        in: query
        name: created_after
        type: string
      - description: 更新时间上限 (RFC3339, e.g. 2025-06-30T23:59:59Z)
        in: query
        name: updated_before
        type: string
      - description: 稀疏字段集，逗号分隔 (e.g. id,name,state,priority)。不传返回完整字段。
        in: query
        name: fields
        type: string
      - description: 排序字段 (-updated_at, priority, target_date, created_at)
        in: query
        name: sort
        type: string
      - description: 每页数量 (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 时间参数格式错误
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 列出工作项
      tags:
      - issue
    post:
      consumes:
      - application/json
      description: 在工作空间中创建需求/任务/缺陷
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/issue.Issue'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 创建工作项
      tags:
      - issue
  /issues/{issue_id}:
    delete:
      responses:
        "204":
          description: No Content
      summary: 删除工作项（归档）
      tags:
      - issue
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Issue'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 获取工作项详情
      tags:
      - issue
    patch:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Issue'
        "409":
          description: 版本冲突
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 更新工作项
      tags:
      - issue
...
  /labels:
    get:
      parameters:
      - description: 状态过滤
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出标签
      tags:
      - label
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/issue.Label'
      summary: 创建标签
      tags:
      - label
  /labels/{label_id}:
    delete:
      parameters:
      - description: 标签 ID
        in: path
        name: label_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除标签
      tags:
      - label
    get:
      parameters:
      - description: 标签 ID
        in: path
        name: label_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Label'
      summary: 获取标签详情
      tags:
      - label
    patch:
      consumes:
      - application/json
      parameters:
      - description: 标签 ID
        in: path
        name: label_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Label'
      summary: 更新标签
      tags:
      - label
...
  /modules:
    get:
      parameters:
      - description: 状态筛选 (active|completed|archived)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/issue.Module'
            type: array
      summary: 列出项目模块
      tags:
      - module
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/issue.Module'
      summary: 创建模块
      tags:
      - module
  /modules/{module_id}:
    delete:
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除模块
      tags:
      - module
    get:
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Module'
      summary: 获取模块详情
      tags:
      - module
    patch:
      consumes:
      - application/json
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issue.Module'
      summary: 更新模块
      tags:
      - module
  /modules/{module_id}/issues:
    get:
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出模块工作项
      tags:
      - module
    post:
      consumes:
      - application/json
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 分配工作项到模块
      tags:
      - module
  /modules/{module_id}/issues/{issue_id}:
    delete:
      parameters:
      - description: 模块 ID
        in: path
        name: module_id
        required: true
        type: integer
      - description: 工作项 ID
        in: path
        name: issue_id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 从模块移除工作项
      tags:
      - module
...
  /public/intake/issues:
    post:
      consumes:
      - application/json
      description: 免登录提交匿名提报工单（无需 Authorization）
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/intake.IntakeIssue'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 公开提交匿名工单
      tags:
      - intake-public
...
securityDefinitions:
  Bearer:
    description: '"格式: Bearer {access_token}。登录/注册接口返回 access_token 后填入。"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"